Refactor: Move miner fee methods to MinerFeeValidation
